A lot of updates

+++ backend/src/utils/pace_calculations.py
@@ -22,7 +22,7 @@ def get_VDOT(dist: float, time: float) -> float:
     vdot = vo2 / percent_max
     return vdot
 
-def get_training_pace_helper(race_dist: int, race_time: int, pct_pace: float) -> int:
+def get_training_pace_helper(race_dist: float, race_time: float, pct_pace: float) -> float:
     vdot = get_VDOT(race_dist, race_time)
     target = round((MILE * 2 * 0.000104)/(-0.182258 + math.sqrt(0.182258**2 - 4 * 0.000104*(-4.6 - pct_pace * vdot))) * 60)
     return target
